+++ src/collective/liveblog/browser/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 from .microupdates import AddMicroUpdateView  # noqa
 from .microupdates import DeleteMicroUpdateView  # noqa
+from .microupdates import EditMicroUpdateView  # noqa
 from .update import Update  # noqa
 from .view import View  # noqa